+++ packages/@livestore/common/src/leader-thread/leader-worker-devtools.ts
@@ -1,9 +1,9 @@
-import { Effect, FiberMap, FiberSet, Option, PubSub, Queue, Stream, SubscriptionRef } from '@livestore/utils/effect'
+import { Effect, FiberMap, Option, PubSub, Queue, Stream, SubscriptionRef } from '@livestore/utils/effect'
 
 import { Devtools, IntentionalShutdownCause, liveStoreVersion, UnexpectedError } from '../index.js'
 import { MUTATION_LOG_META_TABLE, SCHEMA_META_TABLE, SCHEMA_MUTATIONS_META_TABLE } from '../schema/mod.js'
 import type { ShutdownChannel } from './shutdown-channel.js'
-import type { DevtoolsContextEnabled, DevtoolsOptions, PersistenceInfoPair } from './types.js'
+import type { DevtoolsOptions, PersistenceInfoPair } from './types.js'
 import { LeaderThreadCtx } from './types.js'
 
 type SendMessageToDevtools = (